feat: add 2 new custom permissions
CreateUserOrAdmin check if the request method is 'POST' to create an user. Else, check if the user is authenticated and if user has 'is_superuser' attribute to list all users. IsCriticOrAdminOrReadOnly check if request method is 'GET' to list all reviews of a specific movie. Else, check if user is authenticated and has 'is_superuser' or 'is_critic' attribute to create a review.